os/board/rtl8730e/rtl8730e_mipi_lcdc: Change to booting backlight 0
For PWM control of the backlight of the ST7785 and ST7701SN LCDs, the maximum brightness is achieved when the PWM output is 0.
Therefore, after booting and initializing the PWM, since the output power is 0, the LCD backlight is set to maximum brightness.

Since we only set the LCD backlight to maximum brightness in cases other than silent reboots,
to set the default backlight brightness to 0 after booting, we need to set the PWM output power to 100 (MAX) after booting.

Therefore, Add to set lcd backlight 0 during lcd init.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>