os/driver/input/ist415: Fix i2c read stuck issue during IC power off.
During the execution of the operation to turn off the TOUCH IC power, if a touch IRQ occurs
before the IRQ disable, the TOUCH IC power may be cut off before the I2C read operation is completed,
resulting in I2C communication failure.

To resolve this, Add mutex to prevent the TOUCH IC read operation and the power OFF operation from operating simultaneously,
and if the power is already turned off before reading the touch data, the read operation is not performed.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>