os/kernel/task: Fix ASSERT fail in mq_send due to task terminate
When a task is terminated we do below things in a sequence
1. Remove the task from the corresponding tasklist (in current issue,
the task is on the mq wait tasklist)
2. Reset the values in the tcb in task_restore API (One of the values
being reset is used to indicate if the task is on the mq wait tasklist)

However, sine the above points are not performed as a single atomic step,
the state of the tcb becomes inconsistent with the tasklist. Due to this,
the mq send api is resulting in ASSERT failure.

To fix this issue, we call task_recover before remove the task from the
tasklist.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>