arch/armv7-a: Fix hang issue while gating and then pausing the CPU
If we gate the CPU and then if we try to pause the same CPU,
we faced the hang issue. This patch resolves the hang issue by
adding the CPU gating check while proccessing the CPU pause
request.

Below are the 4 cases for cpu pause and gating on each core.

1. cpu1 is in cpu pause, cpu 0 gets an assert, calls cpu pause again for cpu1
2. cpu1 calls cpu pause for cpu0, at that time, cpu0 gets assert.
3. cpu1 is in cpu gating, cpu 0 gets an assert, calls cpu gating or pause for cpu 1
4. cpu1 calls cpu gating for cpu0, at that time cpu0 get assert.

This commit is only for case 3.
Case 1 is already handled in our code.
For case 2 and 4, we are facing a hang. we will handle them with new PR later.

Signed-off-by: neel-samsung <neel.patel@samsung.com>