os/kernel/pthread: Pthread mutex wait must not reutrn EINTR
According to posix standard, if a pthread is waiting to aquire
mutex and gets interrupted by a signal, then it is supposed to
execute the signal and then continue to wait for the mutex.

Previously, our code allowed to pass an argument which decided
whether an EINTR error is retured or not. Since this is not
posix compliant behavior, this commit removes this argument.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>