lib/libc/semaphore: Add flag during init to indicate mutex operation
Add the SIGSEM flag always during init if the semaphore is being
initialized with a count of 1. This indicates that the semaphore is
used either as a mutex or for signalling between tasks. This flag will
later be used to make sure that the semcount never goes above 1 for
such semaphores.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>