build/tools/amebasmart/gnu_utility: Modify AES GCM buffer allocation
-Modify AES GCM handle input and output method
-Do not allocate AES GCM input and output, use pass-in buffer

Prior to this PR, AES GCM would allocate 2 secure buffers to handle encryption or decryption.
These buffers would hold the plaintext and ciphertext during encryption or decryption.
It would copy the data from the non-secure incoming buffers and save it.
Once encryption or decryption is complete, the output is copied to the incoming buffers.
This PR removes the allocation of secure buffers and SE will use the incoming buffers directly.