dbuild, build/configs: Add device selection menu
When we download, we either predefined a default port for each board or manually typed
commands like ./dbuild.sh download port=ttyACM0 all.

However, depending on the connection state of the computer, the board might be detected
on a port other than the default one.

Therefore, instead of hardcoding the default port in the script, we will modify it to display
and allow the user to select the connected device port every time during the download process.

like below:
==================================================
  "Port Selected:"
==================================================
  "1. ttyACM0"
  "2. ttyUSB0"
  "3. ttyUSB1"
  "x. Exit"
==================================================
2

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>