framework/src/media: create FocusManager object.
create FocusManager object during FocusManagerWorker creation.
This is because, if FocusManager object is created in application
task, FocusManager object will be destructed when application exits.
This may even lead to crash if some FocusManager API is executing in any thread.
So, create in main_task for FocusManager object to be there for lifetime.