apps/examples/testcase/le_tc/kernel: Fix failing kernel testcases
On running kernel_tc following testcases are failing-
[tc_pthread_set_get_affinity] FAIL [Line : 1641] pthread create : Values (ret_chk == 0x0) and (22 == 0x16) are not equal
[tc_task_task_restart] FAIL [Line : 292] task_restart : Values (ret_chk == 0xffffffff) and (0 == 0x0) are not equal
[tc_task_task_reparent] FAIL [Line : 531] task_reparent : Values (tc_reparent_chk == 0x2) and (0 == 0x0) are not equal
[tc_libc_signal_sigwait] FAIL [Line : 425] waitpid : Values (recv_ret == 0xffffffff) and (ERROR == 0xffffffff) are equal

tc_pthread_set_get_affinity-- For pthread creation in non-smp case only warning will print on passing non zero affinity value, so no error will be returned. Hence our TC must check return value of OK instead of EINVAL.
tc_task_task_restart-- The restart_task() gets executed as soon as task is created and completes before task_restart is called so no tcb is found corresponding to that pid.Thus to prevent restart_task() to finish before restart is called the changes are made.
tc_task_task_reparent-- If a file is opened in a parent pthread, it can directly be used in child pthread, because both threads belong to the same thread group and hence the fd's are duplicated. However, this is not the same with tasks. Each task belongs to its own thread group. So, files opened in parent task cannot be used in child task. Hence we need to again open the testcase driver file in our child task.
tc_libc_signal_sigwait-- When SMP is enabled, Error code returned by waitpid is ECHILD. This is because, it is possible that the signal sender thread and receiver thread are running on different cores. So, receiver is exiting before sender can send a signal. To avoid this, we make the receiver thread to wait and a semaphore which will be posted only after the sender thread has completed sending the signal.

Signed-off-by: Sambhavi <sambhavi.1@samsung.com>