drivers/input: ist415: Register PM domain before enabling IRQ
The ist415 driver calls pm_timedsuspend upon an IRQ, but an error occurs
because the IRQ is enabled before the ist415 PM domain is registered.

This commit ensures that the PM domain is registered before the IRQ is enabled.
The ist415_reset function calls ist415_power_on, which in turn calls
ist415_enable to activate the IRQ. By moving the pm_domain_register call
before the ist415_reset call in the ist415_initialize function, we guarantee
that the PM domain is set up before any IRQ can be handled, preventing the
potential error.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>