os/drivers/lcd: Refactor lcd driver architecture
[THIS IS ONGOING REFACTORING]

The ultimate goal of this change is to create a scalable architecture.
To achieve this, separating the MIPI-DSI, LCD panel, and LCD driver modules.
Implement initialization functions for each panel and register the necessary functions as ops.

The driver operates without knowing the panel and LCD protocol,
mipi_lcd.c provides common functions (lcd_dev_s) for LCDs that use MIPI,
st7785.c provides functions for the panel.

Signed-off-by: seokhun-eom <seokhun.eom@samsung.com>