os/arch/arm/src/amebasmart: fix tdm page size calculation
Previously, 1 entire page size is requested for each DMA channel, even when less than page size (N) bytes are requested

Initially it was expected that WCLK will stop after page is filled up to N bytes, however it keeps running until entire page is filled

The behavior now is that the page is also created up to N bytes. If 2 channels are used, then each page is N/2 size, as they will be combined into 1 output buffer

WCLK is now observed to be only as long as N bytes.