os/kernel/sched : scheduler migration changes
When a CPU is chosen to be halted/hotplugged, the tasks present in
its assigned_task_list are removed and shifted to the global ready_to_run_list for migration.

For selecting a new CPU for the tasks which are being migrated,
First, we try to assign a CPU which is in tasks's affinity.
But if we are unable to find any such CPUs,
we try to assign it any of the available CPUs.